+++ src/Paramecium/Paramecium/Engine/AnimalVision.cs
@@ -10,7 +10,7 @@ namespace Paramecium.Engine
     {
         public static AnimalVisionOutput Observe(Double2d originPosition, double angle, long originId, long originSpeciesId, int frontViewRange, int frontViewRayCount, double frontViewAngleRange, int backViewRange, int backViewRayCount, double backViewAngleRange)
         {
-            if (g_Soup is null || !g_Soup.Initialized) throw new InvalidOperationException("The soup has not been created or initialized.");
+            if (g_Soup is null || !g_Soup.Initialized) throw new SoupNotCreatedOrInitializedException();
 
             int soupSizeX = g_Soup.Settings.SizeX;
             int soupSizeY = g_Soup.Settings.SizeY;